requirements.txt
flask
flask-cors
pyjwt
bcrypt
ccxt
pandas
numpy
joblib
pywebpush
cryptography
tinydb
apscheduler
